[refactor] LottoStore 객체 생성하여, 구입 금액, 로또 티켓 생성 책임 부여
