+++ src/main/java/lotto/domain/LottoStore.java
@@ -1,0 +1,51 @@
+package lotto.domain;
+
+import lotto.view.InputView;
+import lotto.view.OutputView;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public class LottoStore {
+    private static final int MAX_LOTTO_BALL_COUNT = 6;
+
+    private PurchaseAmount purchaseAmount;
+    private LottoTickets lottoTickets;
+
+    public LottoStore() {
+        startInputPurchaseAmount();
+        generateLottoTickets();
+    }
+
+    private void startInputPurchaseAmount() {
+        OutputView.printStartGuide();
+        purchaseAmount = InputView.inputPurchaseAmount();
+        OutputView.printLottePieces(purchaseAmount.lottoTicket());
+        OutputView.printChangeMoney(purchaseAmount.giveChangeMoney());
+    }
+
+    private LottoTicket generateLottoTicket() {
+        List<LottoBall> lottoTicket = new ArrayList<>();
+        for (int i = 0; i < MAX_LOTTO_BALL_COUNT; i++) {
+            lottoTicket.add(LottoBallFactory.getInstance().get(i));
+        }
+        return new LottoTicket(lottoTicket);
+    }
+
+    private void generateLottoTickets() {
+        for (int i = 0; i < purchaseAmount.lottoTicket(); i++) {
+            Collections.shuffle(LottoBallFactory.getInstance());
+            lottoTickets = new LottoTickets(generateLottoTicket());
+        }
+        OutputView.printLottoTicket(lottoTickets);
+    }
+
+    public PurchaseAmount getPurchaseAmount() {
+        return purchaseAmount;
+    }
+
+    public LottoTickets getLottoTickets() {
+        return lottoTickets;
+    }
+}
